Move computed buffer back into model
